fix: Fix overflow on 32-bit platforms when writing big files.
Not all definitions are changed, because for example the `Read` and
`Write` calls are really dealing with `size_t` since they are limited
by what you can keep in memory. While it wouldn't be wrong to use
`int64_t` in these cases, they are a deeply entrenched API which is
also used in the client, so I'd rather not change those.

Changelog: Title
Ticket: MEN-7613

(cherry picked from commit b5b2531668d43952074576459614241bf97cdab4)

+++ libflash/optimized_writer.hpp
@@ -26,16 +26,16 @@ public:
 		io::FileReader &reader,
 		io::FileReadWriterSeeker &writer,
 		size_t blockSize = 1024 * 1024,
-		size_t volumeSize = 0);
+		int64_t volumeSize = 0);
 	common::error::Error Copy(bool optimized);
 
 	void PrintStatistics() const;
 
 	struct Statistics {
-		size_t blocksWritten_ {0};
-		size_t blocksOmitted_ {0};
-		size_t bytesWritten_ {0};
-		size_t bytesTotal_ {0};
+		int64_t blocksWritten_ {0};
+		int64_t blocksOmitted_ {0};
+		int64_t bytesWritten_ {0};
+		int64_t bytesTotal_ {0};
 	};
 
 	const Statistics &GetStatistics() const {
@@ -46,7 +46,7 @@ private:
 	size_t blockSize_;
 	io::FileReader &reader_;
 	io::FileReadWriterSeeker &readWriter_;
-	size_t volumeSize_ {0};
+	int64_t volumeSize_ {0};
 
 	Statistics statistics_;
 };
